feat(electron): Phase 1 Lane A — Floating window as signed-out OAuth front door
Strangler-fig step toward retiring the Electron main window. Main stays alive;
this makes the Floating window self-sufficient for auth so it can be the visible
surface a signed-out user signs in from.

- proxy.ts: un-protect /floating-navigator (DD1) — signed-out card vs signed-in
  navigator render on the SAME route via useUser(); no /login redirect, no /home
  hard-nav after sign-in (Clerk re-renders in place).
- electron-auth-provider: drop the post-setActive /home navigation — no Electron
  window targets /home anymore.
- preload: extract the auth+oauth bridges to a shared factory
  (preload-shared/auth-oauth-bridge) and expose the scoped {auth,oauth} slice in
  the Floating preload (T2) so ElectronAuthProvider activates there; the scope
  keeps ElectronStartupSync a no-op (its method guards fail).
- OAuthManager/main: source the OAuth origin from WindowManager.getWebAppOrigin()
  (window-agnostic, T4); push the one-time sign-in ticket to the INITIATING
  window only (T7, single-recipient → no cross-window double-consume) with a PULL
  fallback + main-window fallback.
- FloatingNavigatorContainer: signed-out branch renders SignedOutFloatingCard.

Tests: electron 308 pass; OAuthManager URL tests retargeted to the delegation;
affected renderer tests green. typecheck + lint clean.

Not yet verified (follow-up): the in-place card->navigator flip on real sign-in
(spike) and the packaged native OAuth round-trip (T20 ship gate).
