+++ electron/main.ts
@@ -1945,7 +1945,7 @@ function setupIPCHandlers(): void {
   // OAuth IPC handlers for browser-based OAuth flows
   // OAuth IPC handlers (Zod-validated)
   // Used when WebView OAuth is blocked (e.g., Google OAuth)
-  typedHandle('oauth-start', async (_event, provider) => {
+  typedHandle('oauth-start', async (event, provider) => {
     try {
       const oauth = ensureOAuthManager()
       if (!oauth) {
@@ -1954,7 +1954,9 @@ function setupIPCHandlers(): void {
       if (!oauth.isProviderSupported(provider)) {
         throw new Error(`Unsupported OAuth provider: ${provider}`)
       }
-      return await oauth.startOAuthFlow(provider)
+      // Pass the initiating renderer so the resulting one-time sign-in ticket is
+      // pushed back to THIS window only (single recipient → no double-consume).
+      return await oauth.startOAuthFlow(provider, event.sender)
     } catch (error) {
       log.error('Failed to start OAuth flow:', error)
       return {
